actually maintain uniqueness in package index
Before, it would've added a package unless it was a duplicate of the first one in the array.

In JavaScript, the "double-tilde idiom" produces results similar to the floor function, but not quite.  It also produces bewilderment in the mind of a programmer unfamiliar with JS minutiae.  It's best to avoid it.