Update test for html output to ignore line endings
On windows line endings are crlf while linux e.a. use lf only. Depending
on the settings of git, line endings of the file expectedOutput.html will
reflect these differences. This will break the e2e test.
As a fix, line endings are normalized using the newly added package 'eol'
before the results of the html test are compared.